+++ polaris-react/src/components/Card/tests/Card.test.tsx
@@ -61,7 +61,7 @@ describe('Card', () => {
     );
 
     expect(card).toContainReactComponent(ShadowBevel, {
-      borderRadius: '3',
+      borderRadius: '300',
     });
   });
 });