+++ polaris-react/src/components/IndexFilters/components/SortButton/SortButton.tsx
@@ -123,11 +123,11 @@ export function SortButton({
     >
       <Box
         minWidth="148px"
-        paddingInlineStart="3"
-        paddingInlineEnd="3"
-        paddingBlockStart="2"
-        paddingBlockEnd="2"
-        borderBlockEndWidth="1"
+        paddingInlineStart="300"
+        paddingInlineEnd="300"
+        paddingBlockStart="200"
+        paddingBlockEnd="200"
+        borderBlockEndWidth="025"
         borderColor="border-subdued"
       >
         <ChoiceList
@@ -138,10 +138,10 @@ export function SortButton({
         />
       </Box>
       <Box
-        paddingInlineStart="1_5-experimental"
-        paddingInlineEnd="1_5-experimental"
-        paddingBlockStart="2"
-        paddingBlockEnd="2"
+        paddingInlineStart="150"
+        paddingInlineEnd="150"
+        paddingBlockStart="200"
+        paddingBlockEnd="200"
       >
         <DirectionButton
           direction="asc"