+++ polaris.shopify.com/pages/examples/popover-with-searchable-listbox.tsx
@@ -305,9 +305,9 @@ function PopoverWithSearchableListboxExample() {
                 position: 'relative',
                 width: '310px',
                 height: '292px',
-                padding: 'var(--p-space-2) 0',
-                borderBottomLeftRadius: 'var(--p-border-radius-2)',
-                borderBottomRightRadius: 'var(--p-border-radius-2)',
+                padding: 'var(--p-space-200) 0',
+                borderBottomLeftRadius: 'var(--p-border-radius-200)',
+                borderBottomRightRadius: 'var(--p-border-radius-200)',
               }}
               onScrolledToBottom={handleLazyLoadSegments}
             >